[Back] modif entités + Dto

+++ Sources/gestion-culturelle-back/src/main/java/com/afpa/cda/service/impl/AnimationServiceImpl.java
@@ -115,8 +115,8 @@ public class AnimationServiceImpl implements IAnimationService {
 			salleDto = modelMapper.map(salle,SalleDto.class);
 		}
 
-		manifDto.setCout(animDto.getPrix()+(duree* salleDto.getFraisJournalier()));
-		manifDto.setPrixBillet((manifDto.getCout()/salleDto.getCapacite())*0.8);
+		manifDto.setCout((int) (animDto.getPrix()+(duree* salleDto.getFraisJournalier())));
+		manifDto.setPrixBillet((int) ((manifDto.getCout()/salleDto.getCapacite())*0.8));
 
 		return manifDto;
 	}
