+++ Sources/gestion-culturelle-back/src/main/java/com/afpa/cda/service/impl/ManifestationServiceImpl.java
@@ -182,8 +182,8 @@ public class ManifestationServiceImpl implements IManifestationService {
 			salleDto = modelMapper.map(salle,SalleDto.class);
 		}
 
-		manifDto.setCout(animDto.getPrix()+(duree* salleDto.getFraisJournalier()));
-		manifDto.setPrixBillet((manifDto.getCout()/salleDto.getCapacite())*0.8);
+		manifDto.setCout((int) (animDto.getPrix()+(duree* salleDto.getFraisJournalier())));
+		manifDto.setPrixBillet((int) ((manifDto.getCout()/salleDto.getCapacite())*0.8));
 		
 		return manifDto;
 	}
